# Enable vectorization support for transpose mlir emitter